requirements.txt
# Core
numpy
pandas
PyYAML

# Deep Learning
torch
torchvision
torchaudio
facenet-pytorch
pillow

# Video & Audio Processing
opencv-python
moviepy
ffmpeg-python

# NLP & ASR
openai-whisper
transformers
sentencepiece
whisper-timestamped

# Face Detection
mtcnn

# Plotting / Visualization
matplotlib
seaborn

# Utility
tqdm
scikit-learn
scipy

# Optional
jupyter
ipykernel
